Upgrade MRI to version 0.2-20120601.1
This alpha build has been updated to work with the newer GNU ARM Embedded
toolchain and the mri.ar included in this update was actually compiled with
that toolchain.  It has however been tested to make sure that it will still
work with the Smoothie project which uses a slighly older version of the
Code Sourcery toolchain.  One good consequence of this change is that the
register context sent by MRI on debug events is 100 bytes smaller since it no
longer has to fake sending floating point registers which don't exist on the
Cortex-M3.

This update also includes changes that were made to allow MRI to work with
bare bone projects that don't link with the mbed SDK libraries.