Make checksum macros work in non-optimized builds.
Previously the checksum macros would bloat the code because the inline math to
calculate the checksums were not constant folded in -O0 builds.  This commit
modifies the CHECKSUM() macro to just call get_checksum() instead in such
builds.

The fact that checksums are no longer constant expressions in non-optimized
builds means that they can't be used as case labels in switch statements. I
modified the affected switch statements to use if / else if statements instead.