Easier to configure MRI in makefile.
The default mode for MRI's usage in Smoothie is the same as before
where it only executes if a crash is encountered and isn't really
useable for things like single stepping, etc.

Now you can change the ENABLE_DEBUG_MONITOR make variable from 0 to 1
and it will switch MRI into a fuller debug experience where it takes
over UART0 and stops at init so that you have an opportunity to set
any breakpoints you might be interested in.