Remove iprintf() call from spi.cpp
People were noticing that MRI was halting and waiting for GDB to attach
immediately at startup when configured for full debugging mode.  This
was because output was being sent to stdout and MRI was waiting to
redirect it to GDB's console.  I tracked this down to an
iprintf("\n") call in spi.cpp so I commented it out like the rest of
the iprintf() calls in this file.