Add maxstacksize command for GDB.
Fills the area of SRAM between the top of the .bss section and the
top of the stack.  Can be used later with the maxstacksize command that
I added to .gdbinit.  This command searches the area between the
heap and stack to find where the fill has been overwritten and assumes
that the stack has grown into that area at some point in time.  This is
used to calculate the maximum stack size used.
NOTE: If your heap does ever shrink, then the maxstacksize command will
      be confused.  However, this has been a pretty rare case for
      Smoothie recently.

Also added these other commands:
heapsize - Displays the current heap size.
stacksize - Displays the current stack size.