Fix for make -j (n>1).
Enforce mbed SDK library dependency in root makefile.  Before it would
try to build both the mbed SDK and the Smoothie firmware at the same
time.

A better solution would be to pull the mbed SDK make steps into the
main Smoothie makefile so that the dependencies were clearer to make
which would allow it to build the SDK in parallel to the Smoothie
firmware.  I don't currently do this as the current build system is
closer to what was used before but now there are just sources and
makefiles which can build the mbed SDK (libs and headers) for us.  I
also expect most people to do their clearn builds from within the src/
directory once they have the mbed SDK built since they don't really
need to even keep looking at that part of the code anymore.