Fix mbed build on Windows machines with Cygwin
I had fixed this awhile ago in the core gcc4mbed makefiles but it
somehow got missed in the makefile used specifically for building the
mbed SDK library in Smoothie.  This fix makes sure that GNU make uses
the Windows Command Prompt instead of the Cygwin bash shell.  Without
this change, you could see errors like:
    /usr/bin/sh: copy: command not found
when building on a Windows machine.