Finish laser coolant flow sensor support
Input must be on port 0 or 2, as it uses the edge detect features that
are only on those ports.  It would be a good idea to include a RC filter
on the input, just like the endstop inputs have - unfortunately, all the
endstops are on port 1, so you can't just use one of those.