First working zheightmap strategy for ZProbe will rpbe 25 points on the pronting surface, create a surface map, and store in "grid" on the SD card.
It uses bilinear interpolation to have Z follow the bed surface throughout
printing the object, and is critical on oversized and axial robot systems.

The module assumes a squared printing surface as it does not change the
plane of the printing surface, but will allow printing on a slightly skew bed.