Migrated a large amount (over 1K) of arrays to AHB0. This has been a major undertaking! This change should make it possible for people to have more modules loaded, like ethernet and a display. (Not that I have either to test.)
I tried doing a 7x7 grid, and it didn't crash, but it does look like there may be a pointer corruption issue. I'm looking into this.

Disabled code that shuffles the order in which variables are annealed. I thought it would be good to de-bias that out of the loop, but it seems that some bias is actually good!