Improved ZGridStrategy: (1) Removed a redundant and incorrect move in the Z-probe.  If cal_offset and/or probe_offset non-zero, then the calibration loop first moves the extruder to the grid position, then makes the correct move.  The first (incorrect) move has been removed.
(2) Corrected the getZOffset() routine, which would incorrectly
compute xIndex and yIndex if cal_offset_x or cal_offset_y is
negative ((int) is not the floor of a negative number).

(3) Improved getZOffset() by extrapolating points outside the
grid using the nearest grid square, such that for purposes of
correction, the plane of each outermost grid square effectively
extends to infinity.  The previous code returned zero for all
areas outside the grid, creating a discontinuity at the grid
boundary.

These changes were made especially for using the ZGridStrategy
with a cartesian printer and the recently popular inductive
proximity sensor.  The proximity sensor must have a metal surface
underneath at all times, so the probe calibration must be set to
a smaller area than the bed plate.  The cal_offset_x and
cal_offset_y plus the bed_x and bed_y serve to create an interior
area for the probe grid.  The new getZOffset() will extrapolate
values to first order from the probe grid out to the edge of the
bed plate.

Changes should not interfere with any existing use of
ZGridStrategy.