Changes for rotary delta homing...
The alpha_max, beta_max, gamma_max are the angle that the actuators are at when homed.

home_offsets (set with M206) are the theta offsets for the homing angle.

When a rotary delta homes it uses the angle given in the alpha_max+home_offset[0] and uses the FK
(actuator_to_cartesian) to input the angles and get the cartesian XYZ for those angles, (added a
new call Robot::reset_actuator_position(a, b, c) where a b c are the theta angles for each actuator).

M306 and trim (M666) is disabled for rotary delta. (unless someone can explain to me how trim would
actually work in this case).