Home sequence adjustemnt
Retract fast only if not delta
Backoff Use the slowest of slow rates for delta
Backoff use the slowest of fast rates if not delta
Slow seek uses slowest of slow rates
move to origin at whatever G0 feedrate is set at as this should always
be acceptable.  If G0 was never set with an F parameter then
default_seek_rate is used
back off endstops first when moving to origin in case all axis do not
need to move, it's never a good idea to have any axis barely on the edge
of a limit trigger as machine vibrations could false trip it.