"Fix" problem with flushing 64 byte packets
When exactly 64 bytes are sent, it isn't flushed immediately.
(it is sent later, whenever we have more bytes to send)
If the data is 63 or less bytes, then it is flushed immediately.

This is a problem if the user is waiting for an immediate response from Smoothie.
For example, when sending the '?' character to get the location.

This "fix" is a work-around: we split the data into packets
of 63 bytes or less.  The actual bug should probably be hunted down
and properly fixed.

Other people have hit this problem.  E.g.
   https://community.nxp.com/message/360747