grbl mode homing response
Added 'ok' response for '$H' command.  As far as I can tell, grbl sends back an ok after a homing cycle.  Currently, smoothie was not (when using the '$H" command), which was creating issues with bCNC having any command immediately following homing commands to be stuck in the send buffer.