change step_scale_factor to step_frequency
change the mm per segment to be specified as the number of segments per
second. The mm per segment is now calculated by speed/freq. This means
the distance per segment changes as the jogger moves faster, taking
longer steps to ensure it doesn't run out of things to do before the
next command.