Junction deviation tweak
In the original code, junction deviation was not calculated at values less than 0.95 (~18.2 degrees).  Any gcode file that allows the machine to reach max speed in a straight line and then goes into a turn consisting of line segments less than 18 degrees of change between them will keep the machine at full speed throughout the turn.  If 10 very short segments are placed together in this manner, the machine can complete a 180 degree movement change at max set speed without decelerating through the turn.  The 0.95 value has been changed to 1.0 allowing the max junction speed calculation to work with any line segments that are not co-linear.  This makes the code match the description in the comment.

This also fixes an issue described in the smoothie forums: http://forum.smoothieware.org/forum/t-2146070/smoothie-takes-a-corner-too-fast.