Optimize Code for Slaving stepper motors. I created a StepperMotorSlave object that inherits StepperMotor object. This new objects only re-implements functions that are using SLave pins.
Inside Robot and extruder objects, we instantiate Slave pins only when configuration is defined to save memory.
When slave pins are instantiated we use a StepperMotorSlave object instead of a standard StepperMotor object. This way we impact Stepping timing only for the actuator linked to a slave. Other actuators won't be impacted.