Add Marlin compatible M109 Rxxx support
Support set-and-wait of an exact temperature as per the Marlin implementation.

If both S and R are specified then the temperature specified by the S parameter
takes precedence.

When processing M109 R<T> the current temperature is used to determine if we
need to heat or cool.  If the current temperature is less than T then it is
equivalent to M109 S<T>.  If the current temperature is higher than T then
we want until the temperature drops to T.