CartGridStrategy: compensate Z scale based on target Z, not current Z
The strategy scales the grid compensation between dampening_start and height_limit,
but the current Z position is used to generate the scale. This can create a discontinuity
when switching layers or Z-hopping because the new position is scaled based on the
previous position-- it shouldn't use any information about where it _was_ to adjust
the new Z height.