Added new Watch Screen for RRDGLCD.
Added drawing functions for RRDGLCD.
Created a new Watch screen. "Fan Speed" is added to the "Prepare" menu to adjust the speed of the cooler. On the Watch screen displays the speed of the cooler in percent. In bitmaps.h added icons for the new Watch screen. Tested on Full Graphic smart controller 12864 (ST7920).