Fix Probe when in G20 mode to use coordinates specified in inches.
G38.2 - G28.5 used to properly use G20 mode, but when I tried to get to the current edge, I found that they now were not in inches anymore.

G30 was half in inches, half in millimeters.  G30 F  was using the feedrate in mm, but the coordinates were in inches.  G30 Z would set the new value of Z in inches, however the return was trying to go to the mm return value... but in inch mode.

This has all been fixed and tested in both G20 and G21